fix(scheduling): move timeline logic to src/lib for real coverage; mirror sonar excludes
Root cause of the SonarCloud new-code coverage failure (40.8%):
- useTimelineModel.ts lived under src/components (vitest-excluded) but Sonar's
  coverage.exclusions only listed src/components/**/*.tsx, so its 286 lines were
  counted as uncovered new code.
- useWeekStaffingSuggestions.ts was extracted verbatim from a coverage-excluded
  component into the measured src/hooks zone, adding untested new lines.

Fix (per lessons [2026-05-16] sonar excludes must mirror vitest):
- Move pure timeline logic to src/lib/timelineModel.ts (measured) — now 100%
  lines / 88.9% branches; thin useMemo hook re-exports from it.
- sonar.coverage.exclusions: add src/components/**/*.ts (mirror vitest src/components/**)
  and the relocated data-fetch hook.
- vitest coverage.exclude: mirror the data-fetch hook exclusion.
- Add branch tests: groupBy 'none', unassigned-last sort, orphan-employee drop,
  gap-to-last-sample, buildTimelineModel assembly.
