+++ sonar-project.properties
@@ -26,10 +26,11 @@ sonar.coverage.exclusions=\
   **/*.test.tsx,\
   **/tests/**,\
   src/components/**/*.tsx,\
-  src/components/**/index.ts,\
+  src/components/**/*.ts,\
   src/pages/**/*.tsx,\
   src/contexts/**/*.tsx,\
   src/hooks/use*.tsx,\
+  src/hooks/useWeekStaffingSuggestions.ts,\
   src/integrations/**,\
   src/types/**,\
   src/App.tsx,\
